+++ src/Data/Functor/Bind/Class.hs
@@ -10,7 +10,7 @@
 #define MIN_VERSION_semigroups(x,y,z) 1
 #endif
 
-#if __GLASGOW_HASKELL__ >= 702 && __GLASGOW_HASKELL <= 706 && defined(MIN_VERSION_comonad) && !(MIN_VERSION_comonad(3,0,3))
+#if __GLASGOW_HASKELL__ >= 702
 {-# LANGUAGE Trustworthy #-}
 #endif