+++ src/Data/Semigroup/Foldable/Class.hs
@@ -1,5 +1,9 @@
 {-# LANGUAGE CPP, TypeOperators #-}
 
+#if __GLASGOW_HASKELL__ >= 702
+{-# LANGUAGE Trustworthy #-}
+#endif
+
 #ifndef MIN_VERSION_semigroups
 #define MIN_VERSION_semigroups(x,y,z) 0
 #endif