+++ src/Data/Semigroup/Traversable/Class.hs
@@ -1,4 +1,8 @@
 {-# LANGUAGE CPP, TypeOperators #-}
+
+#if __GLASGOW_HASKELL__ >= 702
+{-# LANGUAGE Trustworthy #-}
+#endif
 -----------------------------------------------------------------------------
 -- |
 -- Copyright   :  (C) 2011-2015 Edward Kmett